Move normalize-context to a separate file
